requirements.txt
django-environ
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
sqlparse==0.5.3
djangorestframework-simplejwt
django-redis
flake8